Refactor internal/testhelpers to respective files.
