+++ pkg/internal/testhelpers/jaeger_container.go
@@ -1,0 +1,7 @@
+// This file and its contents are licensed under the Apache License 2.0.
+// Please see the included NOTICE for copyright information and
+// LICENSE for a copy of the license
+
+package testhelpers
+
+func StartJaegerContainer() {}
